+++ src/lib/InstantSearch.ts
@@ -2,15 +2,11 @@ import algoliasearchHelper, { AlgoliaSearchHelper } from 'algoliasearch-helper';
 import { Client as AlgoliaSearchClient } from 'algoliasearch';
 import EventEmitter from 'events';
 import index, { Index } from '../widgets/index/index';
-import RoutingManager from './RoutingManager';
-import simpleStateMapping from './stateMappings/simple';
-import historyRouter from './routers/history';
 import version from './version';
 import createHelpers from './createHelpers';
 import {
   createDocumentationMessageGenerator,
   createDocumentationLink,
-  isPlainObject,
   defer,
   noop,
   warning,
@@ -19,11 +15,11 @@ import {
   InsightsClient as AlgoliaInsightsClient,
   SearchClient,
   Widget,
-  StateMapping,
-  Router,
   UiState,
 } from '../types';
 import hasDetectedInsightsClient from './utils/detect-insights-client';
+import { Middleware, MiddlewareDefinition } from '../middleware';
+import { createRouter, RouterProps } from '../middleware/createRouter';
 
 const withUsage = createDocumentationMessageGenerator({
   name: 'instantsearch',
@@ -32,11 +28,6 @@ const withUsage = createDocumentationMessageGenerator({
 function defaultCreateURL() {
   return '#';
 }
-
-export type Routing<TRouteState = UiState> = {
-  router: Router<TRouteState>;
-  stateMapping: StateMapping<TRouteState>;
-};
 
 /**
  * Global options for an InstantSearch instance.
@@ -106,7 +97,7 @@ export type InstantSearchOptions<TRouteState = UiState> = {
    * Router configuration used to save the UI State into the URL or any other
    * client side persistence. Passing `true` will use the default URL options.
    */
-  routing?: Partial<Routing<TRouteState>> | boolean;
+  routing?: RouterProps<TRouteState> | boolean;
 
   /**
    * the instance of search-insights to use for sending insights events inside
@@ -136,7 +127,7 @@ class InstantSearch extends EventEmitter {
   public _createURL: (nextState: UiState) => string;
   public _searchFunction?: InstantSearchOptions['searchFunction'];
   public _mainHelperSearch?: AlgoliaSearchHelper['search'];
-  private _routingManager?: RoutingManager;
+  public middleware: MiddlewareDefinition[] = [];
 
   public constructor(options: InstantSearchOptions) {
     super();
@@ -158,14 +149,6 @@ class InstantSearch extends EventEmitter {
 
     if (searchClient === null) {
       throw new Error(withUsage('The `searchClient` option is required.'));
-    }
-
-    if (typeof (options as any).urlSync !== 'undefined') {
-      throw new Error(
-        withUsage(
-          'The `urlSync` option was removed in InstantSearch.js 3. You may want to use the `routing` option.'
-        )
-      );
     }
 
     if (typeof (searchClient as any).search !== 'function') {
@@ -246,33 +229,35 @@ See ${createDocumentationLink({
       this._searchFunction = searchFunction;
     }
 
-    const defaultRoutingOptions = {
-      stateMapping: simpleStateMapping(),
-      router: historyRouter(),
-    };
-
-    let routingOptions: Routing | null = null;
-    if (routing === true) {
-      routingOptions = defaultRoutingOptions;
-    } else if (isPlainObject(routing)) {
-      routingOptions = {
-        ...defaultRoutingOptions,
-        ...routing,
-      };
+    if (routing) {
+      const routerOptions = typeof routing === 'boolean' ? undefined : routing;
+      this.EXPERIMENTAL_use(createRouter(routerOptions));
     }
+  }
 
-    if (routingOptions) {
-      this._routingManager = new RoutingManager({
-        ...routingOptions,
-        instantSearchInstance: this,
+  /**
+   * Hooks a middleware into the InstantSearch lifecycle.
+   *
+   * This method is considered as experimental and is subject to change in
+   * minor versions.
+   */
+  public EXPERIMENTAL_use(...middleware: Middleware[]): this {
+    const newMiddlewareList = middleware.map(fn => {
+      const newMiddleware = fn({ instantSearchInstance: this });
+      this.middleware.push(newMiddleware);
+
+      return newMiddleware;
+    });
+
+    // If the instance has already started, we directly subscribe the
+    // middleware so they're notified of changes.
+    if (this.started) {
+      newMiddlewareList.forEach(m => {
+        m.subscribe();
       });
-
-      this._createURL = this._routingManager.createURL;
-      this._initialUiState = {
-        ...initialUiState,
-        ...this._routingManager.read(),
-      };
     }
+
+    return this;
   }
 
   /**
@@ -430,15 +415,15 @@ See ${createDocumentationLink({
 
     this.mainHelper = mainHelper;
 
+    this.middleware.forEach(m => {
+      m.subscribe();
+    });
+
     this.mainIndex.init({
       instantSearchInstance: this,
       parent: null,
       uiState: this._initialUiState,
     });
-
-    if (this._routingManager) {
-      this._routingManager.subscribe();
-    }
 
     mainHelper.search();
 
@@ -477,9 +462,9 @@ See ${createDocumentationLink({
     this.mainHelper = null;
     this.helper = null;
 
-    if (this._routingManager) {
-      this._routingManager.dispose();
-    }
+    this.middleware.forEach(m => {
+      m.unsubscribe();
+    });
   }
 
   public scheduleSearch = defer(() => {
@@ -512,9 +497,11 @@ See ${createDocumentationLink({
   public onStateChange = () => {
     const nextUiState = this.mainIndex.getWidgetState({});
 
-    if (this._routingManager) {
-      this._routingManager.write({ state: nextUiState });
-    }
+    this.middleware.forEach(m => {
+      m.onStateChange({
+        state: nextUiState,
+      });
+    });
   };
 
   public createURL(nextState: UiState = {}): string {