+++ src/middleware/index.ts
@@ -1,0 +1,13 @@
+import { UiState } from '../types';
+
+export interface MiddlewareDefinition {
+  onStateChange(options: { state: UiState }): void;
+  subscribe(): void;
+  unsubscribe(): void;
+}
+
+export type Middleware = ({
+  instantSearchInstance: InstantSearch,
+}) => MiddlewareDefinition;
+
+export { createRouter, RouterProps } from './createRouter';